+++ programs/client/CMakeLists.txt
@@ -2,7 +2,6 @@ set (CLICKHOUSE_CLIENT_SOURCES
     Client.cpp
     ConnectionParameters.cpp
     QueryFuzzer.cpp
-    Suggest.cpp
 )
 
 set (CLICKHOUSE_CLIENT_LINK
